ml-service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
cmdstanpy>=1.2.0
prophet>=1.1.0
